requirements/qa.txt
pytest==3.6.3
pytest-flake8==1.0.2
flake8==3.5.0
pyflakes==1.6.0
pytest-mccabe==0.1
coverage==4.5.1
pytest-cov==2.5.1
requests-mock==1.5.2